@@ -20,6 +20,7 @@ import { DocumentsComponents } from './components/creatnewlead/documents-1/docum
 import { VreportComponent } from './components/vreport/vreport.component';
 import { CommentsComponent } from './components/comments/comments.component';
 import { InvoiceComponent } from './components/invoice/invoice.component';
+import { TablereportComponent } from './components/tablereport/tablereport.component';
 
 const routes: Routes = [
 
@@ -31,17 +32,10 @@ const routes: Routes = [
   {path:'lead',component:LeadComponent},
   {path:'newtask',component:NewtaskComponent},
   {path:'verify',component:VreportComponent},
-  {path:'newtask',component:NewtaskComponent},
-  {path:'newtask',component:NewtaskComponent},
 
-
-// {path:'comments',component:CommentsComponent},
- 
+  {path:'newtask',component:NewtaskComponent},
 
 {path:"invoice",component:InvoiceComponent},
-
-  // {path:'verification',component:VerificationComponent},
-
   {path:'comments',component:CommentsComponent},
   {path:'Overview',component:OverviewComponent},
   {path:"Application-details",component:AplicationDetailsComponent},
@@ -55,12 +49,12 @@ const routes: Routes = [
   {path:'vreport',component:VreportComponent},
   {path:'documents-1',component:DocumentsComponents},
   {path:'verification',component:VerificationComponent},
-
+  {path:'tabelreport',component:TablereportComponent},
 
       // { path: '', redirectTo: '/vreport', pathMatch: 'full' }
-  {path:'',component:VerificationComponent, children:[
-    {path:'vreport',component:VreportComponent},
-    {path:'comments',component:CommentsComponent},
+//   {path:'',component:VerificationComponent, children:[
+//     {path:'vreport',component:VreportComponent},
+//     {path:'comments',component:CommentsComponent},
 
 
   // {path:'',component:VerificationComponent, children:[
@@ -68,14 +62,14 @@ const routes: Routes = [
   // ]}
 
 
-  ]},
+  
+
 
   
-  {path:'vreport',component:VreportComponent},
+
   { path: '', redirectTo: 'verification', pathMatch: 'full' },
   {
-    path: 'verification',
-    component: VerificationComponent,
+    path: 'verification',component: VerificationComponent,
     children: [
       {path:'Overview',component:OverviewComponent},
       {path:"Application-details",component:AplicationDetailsComponent},
@@ -85,37 +79,15 @@ const routes: Routes = [
         // { path: 'comments', component: CommentsComponent },
         // { path: 'vreport', component: VreportComponent },
         {path:'empdoc',component:BrowseempdocumentsComponent},
+      { path: 'comments', component: CommentsComponent },
+      { path: 'vreport', component: VreportComponent },
+      {path:'empdoc',component:BrowseempdocumentsComponent},
       ],
     },
  
 ]
 
-// const routes: Routes = [
-  // { path: 'login', component: LoginComponent },
-  // { path: 'managerlogin', component: ManagerLoginComponent },
-  // { path: 'registration', component: RegistrationComponent },
-  // { path: 'checkeligibility', component: CheckeligibilityComponent },
-  // { path: 'personalloan', component: PersonalloanComponent },
-  // { path: 'lead', component: LeadComponent },
-  // { path: 'newtask', component: NewtaskComponent },
-  // { path: 'verify', component: VreportComponent },
-  // { path: 'empdocuments', component: EmployeeDocumentsComponent },
-  // { path: 'empdoc', component: BrowseempdocumentsComponent },
-  // { path: 'verification', component: VerificationComponent },
-  // { path: 'overview', component: OverviewComponent },
-  // { path: 'application-details', component: AplicationDetailsComponent },
-  // { path: 'loan-details', component: LoanDetailsComponent },
-  // { path: 'bank', component: BankComponent },
-  // { path: '', redirectTo: 'verification', pathMatch: 'full' },
-  // {
-  //   path: 'verification',
-  //   component: VerificationComponent,
-  //   children: [
-  //     { path: 'comments', component: CommentsComponent },
-  //     { path: 'vreport', component: VreportComponent },
-  //   ],
-  // },
-// ];
+
 
 
 @NgModule({
