@@ -7,39 +7,10 @@ import { Component } from '@angular/core';
 })
 export class BrowseempdocumentsComponent {
 
-  // documents = [
-  //   { placeholder: 'Photo' },
-  //   { placeholder: 'Aadhar Card' },
-  //   { placeholder: '3 Months Salary Slip' },
-  //   { placeholder: 'Last 2 Years Form 16' },
-  //   { placeholder: 'Proof Of Bonus Income' },
-  //   { placeholder: 'Property Cost Sheet/ Sale Agreement' },
-  //   { placeholder: 'Index 2' },
-  //   { placeholder: 'Pan Card' },
-  //   { placeholder: 'Address Proof' },
-  //   { placeholder: '6 Months Bank Statement' },
-  //   { placeholder: 'Proof Of Rent Income' },
-  //   { placeholder: 'Proof Of Incentive Income' },
-  //   { placeholder: 'Floor Plan' }
-  // ];
-
-  // handleFileUpload(id: number, event: Event): void {
-  //   const input = event.target as HTMLInputElement;
-  //   const textField = document.getElementById(`file-url-${id}`) as HTMLInputElement;
-
-  //   if (input && textField && input.files && input.files.length > 0) {
-  //     textField.value = input.files[0].name; 
-  //   } else {
-  //     console.error('No file selected or text field not found.');
-  //   }
-  // }
-
-// fileName:String='icon.png';
-
   documents = [
     { placeholder: 'Photo', fileName: '' },
     { placeholder: 'Aadhar Card Front-side', fileName: '' },
-     { placeholder: 'Aadhar Card Back-side', fileName: '' },
+    { placeholder: 'Aadhar Card Back-side', fileName: '' },
     { placeholder: '3 Months Salary Slip', fileName: '' },
     { placeholder: 'Last 2 Years Form 16', fileName: '' },
     { placeholder: 'Proof Of Bonus Income', fileName: '' },
@@ -53,56 +24,130 @@ export class BrowseempdocumentsComponent {
     { placeholder: 'Floor Plan', fileName: '' }
   ];
 
-  isModalVisible: boolean = false;  // To control modal visibility
-  selectedDocument: any = {}; // Store the selected document for viewing
+   documents2 = [
+    { placeholder: 'Passport', fileName: '' },
+    { placeholder: 'Voter ID', fileName: '' },
+    { placeholder: 'Driving License', fileName: '' },
+    { placeholder: 'Bank Passbook', fileName: '' },
+    { placeholder: 'Electricity Bill', fileName: '' },
+    { placeholder: 'Water Bill', fileName: '' },
+    { placeholder: 'Gas Bill', fileName: '' },
+    { placeholder: 'Telephone Bill', fileName: '' },
+    { placeholder: 'Rent Agreement', fileName: '' },
+    { placeholder: 'Vehicle Registration Document', fileName: '' },
+    { placeholder: 'Income Tax Returns (Last 2 Years)', fileName: '' },
+    { placeholder: 'Company ID Card', fileName: '' },
+    { placeholder: 'Educational Certificates', fileName: '' },
+    { placeholder: 'Marriage Certificate', fileName: '' }
+  ];
+  
 
-  // Handle file upload and update the text field with the file name
-  handleFileUpload(index: number, event: Event): void {
-    const input = event.target as HTMLInputElement;
-    if (input && input.files && input.files.length > 0) {
-      const fileName = input.files[0].name;
-      this.documents[index].fileName = fileName;  // Store the file name in the document
+  selectedDocuments: any[] = this.documents; // Default to primary documents
+  isPrimarySelected: boolean = true;  // Initially Primary is selected
+  isModalVisible: boolean = false;
+  selectedDocument: any = {};
+
+  // Function to switch between Primary and Secondary
+  selectApplicant(applicantType: string): void {
+    if (applicantType === 'primary') {
+      this.selectedDocuments = this.documents;
+      this.isPrimarySelected = true;
     } else {
-      console.error('No file selected or input not found.');
+      this.selectedDocuments = this.documents2;
+      this.isPrimarySelected = false;
     }
   }
 
-  // Reset the file and restore the placeholder when X is clicked
+//  handleFileUpload (index: number, event: Event): void {
+//     const input = event.target as HTMLInputElement;
+//     if (input && input.files && input.files.length > 0) {
+//       const file = input.files[0];
+//       const fileName = file.name;
+//       const fileSize = file.size / (1024 * 1024); // Convert size to MB
+//       const fileExtension = fileName.split('.').pop()?.toLowerCase();
+
+//       // Allowable extensions
+//       const allowedExtensions = ['jpg', 'jpeg', 'pdf'];
+
+//       // Check file extension and size
+//       if (!allowedExtensions.includes(fileExtension || '')) {
+//         alert('Only JPG, JPEG, and Pdf files are allowed.');
+//         return;
+//       }
+//       if (fileSize > 2) {
+//         alert('File size must not exceed 2MB.');
+//         return;
+//       }
+
+//       // Valid file: store the file name in the document
+//       this.selectedDocuments[index].fileName = fileName;
+//     } else {
+//       console.error('No file selected or input not found.');
+//     }
+//   }
+
+
+
+handleFileUpload(index: number, event: Event): void {
+  const input = event.target as HTMLInputElement;
+  if (input && input.files && input.files.length > 0) {
+    const file = input.files[0];
+    const fileName = file.name;
+    const fileSize = file.size / (1024 * 1024); // Convert size to MB
+    const fileExtension = fileName.split('.').pop()?.toLowerCase();
+
+    // Allowable extensions
+    const allowedExtensions = ['jpg', 'jpeg', 'png', 'pdf'];
+
+    // Check file extension and size
+    if (!allowedExtensions.includes(fileExtension || '')) {
+      alert('Only JPG, JPEG, PNG, and PDF files are allowed.');
+      return;
+    }
+    if (fileSize > 2) {
+      alert('File size must not exceed 2MB.');
+      return;
+    }
+
+    // Valid file: store the file name in the document
+    this.selectedDocuments[index].fileName = fileName;
+  } else {
+    console.error('No file selected or input not found.');
+  }
+}
+
+
   resetFile(index: number): void {
-    this.documents[index].fileName = '';  // Clear the uploaded file name
+    this.selectedDocuments[index].fileName = '';  // Clear the uploaded file name
   }
 
-  // Handle view document click (open modal)
   viewDocument(index: number): void {
-    this.selectedDocument = this.documents[index];
-    this.isModalVisible = true;
+    const document = this.selectedDocuments[index];
+    if (!document.fileName) {
+      this.selectedDocument = null;
+      this.isModalVisible = true;
+    } else {
+      this.selectedDocument = document;
+      this.isModalVisible = true;
+    }
   }
 
-  // Close the modal
   closeModal(): void {
     this.isModalVisible = false;
+    this.selectedDocument = null;
   }
 
-  // Check if the file is an image
-  // isImage(fileName: string): boolean {
-  //   const imageExtensions = ['.jpg', '.jpeg', '.png', '.gif'];
-  //   const fileExtension = fileName.split('.').pop()?.toLowerCase();
-  //   return imageExtensions.includes(fileExtension || '');
-  // }
   isImage(fileName: string): boolean {
-    const imageExtensions = ['jpg', 'jpeg', 'png', 'gif', 'bmp', 'webp','pdf'];
+    const imageExtensions = ['jpg', 'jpeg', 'pdf'];
     const fileExtension = fileName.split('.').pop()?.toLowerCase();
     return fileExtension ? imageExtensions.includes(fileExtension) : false;
   }
+
   getFileUrl(fileName: string): string {
-    return `${fileName}`; // Replace with your file path logic
+    return `${fileName}`;
   }
+
   onImageError(event: Event): void {
     (event.target as HTMLImageElement).src = 'placeholder.png';
   }
-  
-  // Get the file URL (for image preview)
-  // getFileUrl(fileName: string): string {
-  //   return `/uploads/${fileName}`;  // You can customize this based on your file storage path
-  // }
 }
